requirements.txt
Augmentor
requests 
pillow 
opencv-python